+++ voctocore/lib/sources/avsource.py
@@ -51,7 +51,7 @@ class AVSource(object, metaclass=ABCMeta):
             # check if we have video to show no-signal message
             assert self.has_video
             # set timeout at which we check for signal loss
-            GLib.timeout_add(self.timer_resolution * 1000, self.do_timeout)
+            GLib.timeout_add(int(self.timer_resolution * 1000), self.do_timeout)
         
         # this might get attached to the no-signal compositor's input sink
         self.noSignalSink = None
